refactor: entity id long타입으로 변경

+++ src/main/java/com/health/healthlog/repository/BodyRepository.java
@@ -3,5 +3,5 @@ package com.health.healthlog.repository;
 import com.health.healthlog.domain.Body;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BodyRepository extends JpaRepository<Body, Integer> {
+public interface BodyRepository extends JpaRepository<Body, Long> {
 }
